requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.2
pydantic==1.10.2
PyYAML==6.0
requests==2.28.1
selenium==4.7.2
tenacity==8.1.0
